msm: thermal: simple: Make get_valid_cpufreq() usage hotplug-safe
One of the clusters can be offline when one of the zone sysfs nodes is
written to, causing get_valid_cpufreq() to fail because it cannot read the
cpufreq policy for an offline CPU. Calling get_valid_cpufreq() for an
offline CPU can also potentially crash the kernel due to unsafe
cpufreq_frequency_get_table() usage (it is lacking get/put_online_cpus()
protection). If the kernel does not crash after this, then an incompatible
frequency can be stored inside the thermal zone freq array and create
cpufreq errors.

To avoid all of the mayhem the incorrect get_valid_cpufreq() usage can
cause, only use it inside the cpufreq notifier where hotplug protection
is guaranteed for the CPU that the notifier is executing for.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Joe Maples <joe@frap129.org>